+++ R/datasets_aquarium_mod.R
@@ -8,7 +8,8 @@
 #' The \code{\link{aquarium_run}} dataset is a corresponding MCMC run.
 #' 
 #' @examples
-#' library(tidyverse)
+#' library(tibble)
+#' library(dplyr)
 #' exp <- tibble::tribble(
 #'   ~time.day,  ~species, ~biomass, ~prop15N,
 #'           0,   "algae",     1.02,  0.00384,
@@ -30,8 +31,8 @@
 #'         2.5,     "NH4",       NA,    0.295,
 #'           3,     "NH4",     0.27,        NA
 #'   )
-#' inits <- exp %>% filter(time.day == 0)
-#' obs <- exp %>% filter(time.day > 0)
+#' inits <- exp %>% dplyr::filter(time.day == 0)
+#' obs <- exp %>% dplyr::filter(time.day > 0)
 #' 
 #' aquarium_mod <- new_networkModel() %>%
 #'     set_topo("NH4 -> algae -> daphnia -> NH4") %>%