+++ R/ui_others.R
@@ -24,7 +24,7 @@
 #' Standards recognized by this function are: \code{c("d15N", "d2H", "d13C",
 #' "d17O.SMOW", "d18O.SMOW", "d33S", "d34S", "d36S")}
 #' 
-#' @param x Vector of delta values
+#' @param x Vector of delta values.
 #' @param Rstandard String describing the isotopic measurement, e.g. "d15N",
 #'     "d13C" and used to set automatically Rstandards (see the Section
 #'     "Ratios for reference standards" for more details). Alternatively, a
@@ -73,9 +73,23 @@ delta2prop <- function(x = NULL, Rstandard = NULL) {
     if (is.null(Rstandard)) {
         Rstandard <- "non-specified"
         if (is.null(x)) {
-            cat(msg, "\n")
+            message(paste(msg, sep = "\n"))
             return(invisible(NULL))
         }
+    }
+    if (length(Rstandard) > 1) {
+      msg <- paste0(
+        "The Rstandard argument must be of length 1 but the argument provided was ",
+        ifelse(is.numeric(Rstandard), "a numeric vector", "an object"),
+        " of length ", length(Rstandard), ".")
+      if (is.numeric(Rstandard)) {
+        msg <- paste0(
+          msg, "\n",
+          "If you did not want to pass a numeric vector but rather wanted to ",
+          "pass a string defining the standard to use (such as \"d15N\"), maybe ",
+          "you forgot the quotes?")
+      }
+      stop(msg)
     }
     if (is.character(Rstandard)) {
         if (!Rstandard %in% names(Rstandards)) {
@@ -92,6 +106,95 @@ delta2prop <- function(x = NULL, Rstandard = NULL) {
     R0 <- Rstandard
     props <- R0 * (x/1000 + 1) / (R0 * (x/1000 + 1) + 1)
     return(props)
+}
+
+### * prop2delta()
+
+#' Convert isotopic proportions to delta values
+#'
+#' This function performs the inverse of the operation performed by
+#' \code{delta2prop()}.
+#'
+#' @param x Vector of proportions values.
+#' @param Rstandard String describing the isotopic measurement, e.g. "d15N",
+#'     "d13C" and used to set automatically Rstandards (see the Section
+#'     "Ratios for reference standards" for more details). Alternatively, a
+#'     numeric value to use for Rstandard, e.g. 0.0036765.
+#'
+#' @return A vector of same length of x, containing the delta values based on
+#'   the proportions of heavy isotope provided as x and the Rstandard provided.
+#'
+#' @examples
+#' prop15N <- c(0.00395, 0.02222, 0.00462, 0.00753, NA, 0.00422, 0.00492)
+#'
+#' # Rstandard can be specified with a string for some preset references
+#' d15N <- prop2delta(prop15N, "d15N")
+#' d15N
+#'
+#' # Rstandard can also be specified manually for non-preset references
+#' d15N_manual <- prop2delta(prop15N, 0.0036765)
+#' d15N_manual
+#'
+#' # Call delta2prop() to get the detail of available references
+#' delta2prop()
+#'
+#' @export
+
+prop2delta <- function(x = NULL, Rstandard = NULL) {
+    # Known standards
+    Rstandards <- list("d15N" = 0.0036765,
+                       "d2H" = 0.00015576,
+                       "d13C" = 0.011180,
+                       "d17O.SMOW" = 0.0003799,
+                       "d18O.SMOW" = 0.0020052,
+                       "d33S" = 0.0078772,
+                       "d34S" = 0.0441626,
+                       "d36S" = 0.0001533)
+    # Message about available standards
+    msg <- paste0("Available standards are:")
+    for (i in seq_along(Rstandards)) {
+        label <- names(Rstandards)[i]
+        string <- paste(label, paste(rep("_", 13 - nchar(label)), collapse = ""))
+        value <- Rstandards[[i]]
+        msg <- paste(msg, paste(string, value, collapse = ""), sep = "\n    ")
+    }
+    # Parse argument
+    if (is.null(Rstandard)) {
+        Rstandard <- "non-specified"
+        if (is.null(x)) {
+            message(paste(msg, sep = "\n"))
+            return(invisible(NULL))
+        }
+    }
+    if (length(Rstandard) > 1) {
+      msg <- paste0(
+        "The Rstandard argument must be of length 1 but the argument provided was ",
+        ifelse(is.numeric(Rstandard), "a numeric vector", "an object"),
+        " of length ", length(Rstandard), ".")
+      if (is.numeric(Rstandard)) {
+        msg <- paste0(
+          msg, "\n",
+          "If you did not want to pass a numeric vector but rather wanted to ",
+          "pass a string defining the standard to use (such as \"d15N\"), maybe ",
+          "you forgot the quotes?")
+      }
+      stop(msg)
+    }
+    if (is.character(Rstandard)) {
+        if (!Rstandard %in% names(Rstandards)) {
+            unk_std <- paste0("Rstandard argument (", Rstandard, "): unknown. ",
+                              collapse = "")
+            msg <- paste0(unk_std, msg, collapse = "")
+            stop(msg)
+        }
+        Rstandard <- Rstandards[[Rstandard]]
+    } else if (!is.numeric(Rstandard)) {
+        stop("Provided value must the name of a known measurement type or a numeric.")
+    }
+    # Calculate deltas and return
+    R0 <- Rstandard
+    deltas <- (1/R0 * x / (1 - x) - 1) * 1000
+    return(deltas)
 }
 
 ### * filter_by_group